Fold `TestManagerBase` and related functions into `TestManager`
